Round 1 redesign... GO!
Refactor some methods into a utility class.

Design a utility class for chained literal replacements. Two
implications here:
   ) Some of the old methodology would not replace the literal tokens
when not active. It will now always replace a matched token.
   ) Some of the old methodology would have fallen over if % was used in
a pattern to create a new match for literal. Although this could in rare
circumstances be predictable, it is not appropriate and has been
corrected.

The new style now uses a callback interface to determine replacement
values, with a static API that can efficiently run through the list of
possible literals when the primary token is matched. A 'scope' is
provided as a parameter, storing the ChatPlayer originating the call.